feat: implement src/modules
